Create ingestion package
