+++ rundetection/rules/common_rules.py
@@ -1,7 +1,7 @@
 """
 Module containing rule implementations for instrument shared rules
 """
-from rundetection.ingest import JobRequest
+from rundetection.ingestion.ingest import JobRequest
 from rundetection.rules.rule import Rule
