+++ rundetection/rules/rule.py
@@ -4,7 +4,7 @@ Module containing the abstract base Rule class and MissingRuleError
 from abc import ABC, abstractmethod
 from typing import Generic, TypeVar
 
-from rundetection.ingest import JobRequest
+from rundetection.ingestion.ingest import JobRequest
 
 T = TypeVar("T")
