+++ rundetection/run_detection.py
@@ -15,7 +15,7 @@ from typing import Generator, Any
 from pika import BlockingConnection, ConnectionParameters, PlainCredentials  # type: ignore
 from pika.adapters.blocking_connection import BlockingChannel  # type: ignore
 
-from rundetection.ingest import ingest, JobRequest
+from rundetection.ingestion.ingest import ingest, JobRequest
 from rundetection.specifications import InstrumentSpecification
 
 file_handler = logging.FileHandler(filename="run-detection.log")
