+++ test/rules/test_common_rules.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-from rundetection.ingest import JobRequest
+from rundetection.ingestion.ingest import JobRequest
 from rundetection.rules.common_rules import EnabledRule
